Refactor: remove code duplication, reorganize
