+++ core/src/main/scala/akka/MesmerMirrorTypes.scala
@@ -8,6 +8,7 @@ import akka.stream.stage.GraphStageLogic
 
 object MesmerMirrorTypes {
   type ActorRefWithCell                = akka.actor.ActorRefWithCell
+  type ActorSystemImpl                 = akka.actor.ActorSystemImpl
   type Cell                            = akka.actor.Cell
   type ExtendedActorMaterializerMirror = ExtendedActorMaterializer
   type GraphInterpreterMirror          = GraphInterpreter
