+++ core/src/main/scala/io/scalac/mesmer/core/util/ClassicActorSystemOps.scala
@@ -1,16 +1,16 @@
-package akka
+package io.scalac.mesmer.core.util
 
 import _root_.io.scalac.mesmer.core.invoke.Lookup
+import akka.MesmerMirrorTypes.ActorSystemImpl
 import akka.actor.ActorSystem
-import akka.actor.ActorSystemImpl
 import akka.util.Unsafe
 
-object ActorSystemOps extends Lookup {
+object ClassicActorSystemOps extends Lookup {
 
   private val initializedOffset: Long =
     Unsafe.instance.objectFieldOffset(classOf[ActorSystemImpl].getDeclaredField("_initialized"))
 
-  implicit final class ActorSystemOpsWrapper(private val system: ActorSystem) extends AnyVal {
+  implicit final class ActorSystemOps(private val system: ActorSystem) extends AnyVal {
 
     /**
      * Unsafe is used to ensure volatile semantics on field access
