+++ otel-extension/src/main/scala/io/scalac/mesmer/otelextension/instrumentations/akka/stream/AkkaStreamMonitorExtension.scala
@@ -1,6 +1,5 @@
 package io.scalac.mesmer.otelextension.instrumentations.akka.stream
 
-import akka.ActorSystemOps.ActorSystemOpsWrapper
 import akka.actor.ActorRef
 import akka.actor.typed.ActorSystem
 import akka.actor.typed.Behavior
@@ -21,6 +20,7 @@ import io.scalac.mesmer.core.event.StreamEvent
 import io.scalac.mesmer.core.event.StreamEvent.LastStreamStats
 import io.scalac.mesmer.core.event.StreamEvent.StreamInterpreterStats
 import io.scalac.mesmer.core.model.ShellInfo
+import io.scalac.mesmer.core.util.ClassicActorSystemOps.ActorSystemOps
 import io.scalac.mesmer.core.util.Retry
 import io.scalac.mesmer.core.util.stream
 import io.scalac.mesmer.otelextension.instrumentations.akka.stream.AkkaStreamMonitorExtension.StreamStatsReceived
